[+]Added a authentication with different access

+++ src/main/java/com/mpbtwozeroone/piebook/repositories/RecipeRepository.java
@@ -3,5 +3,5 @@ package com.mpbtwozeroone.piebook.repositories;
 import com.mpbtwozeroone.piebook.model.Recipe;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RecipeRepository extends JpaRepository<Recipe, Integer> {
+public interface RecipeRepository extends JpaRepository<Recipe, Long> {
 }
